change consumption to gallons

+++ db/migrate/20150815185126_create_usages.rb
@@ -2,7 +2,7 @@ class CreateUsages < ActiveRecord::Migration
   def change
     create_table :usages do |t|
       t.string :account_id, null: false
-      t.integer :consumption, null: false
+      t.float :consumption, null: false
       t.string :month, null: false
 
       t.timestamps null: false
